% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Word_to_RMarkdown.R
\name{create_Rmd}
\alias{create_Rmd}
\title{Automatically creates a table of contents for a .Rmd file}
\usage{
create_Rmd(doc, toc = TRUE)
}
\arguments{
\item{doc}{word document string}

\item{toc}{boolean to enable table of contents (toc)}
}
\value{
.Rmd file string
}
\description{
Scans documents for headings (#) and creates a table of contents (hyper linked). The returned string can directly be written as .Rmd file.
All code section will be excluded for the (#) search.
}
\details{
As the function argument pass a R markdown .Rmd file string which can be read via \code{readLines("fileName.Rmd")}.
}
