% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmarkdown_utiilities.R
\name{r_toc_for_Rmd}
\alias{r_toc_for_Rmd}
\title{Automatically creates a table of contents for ".Rmd" file}
\usage{
r_toc_for_Rmd(
  c_Rmd,
  toc_heading_string = "Table of Contents",
  create_nb = TRUE,
  create_top_link = TRUE,
  nb_front = TRUE,
  set_first_heading_level = FALSE
)
}
\arguments{
\item{c_Rmd}{Rmd file string}

\item{toc_heading_string}{title string for TOC, e.g "Inhaltsverzeichnis", default is "Table of contents".}

\item{create_nb}{boolean to enable numbering for each heading.}

\item{create_top_link}{boolean to create link below each heading to jump back to the table of contents.}

\item{nb_front}{boolean to have the numbering in front of each heading.}

\item{set_first_heading_level}{set first found heading level to heading 1}
}
\value{
.Rmd file string
}
\description{
Scans documents for headings (#) and creates a table of contents (hyper linked). The returned string can directly be written as .Rmd file.
All code section will be excluded for the (#) search.It is very similar to the following setting: output:
html_document:
  toc: true
  number_sections: true
}
\details{
Pass a R markdown .Rmd file string as first function argument. Read a file via \code{readLines("fileName.Rmd")}.
}
\examples{
print(tbl_of_contents.Rmd)
c_rmd <- r_toc_for_Rmd(tbl_of_contents.Rmd, create_nb = FALSE)
print(c_rmd)
c_rmd <- r_toc_for_Rmd(tbl_of_contents.Rmd,toc_heading_string = "Inhaltsverzeichnis", create_nb = TRUE)
print(c_rmd)
}
